+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # 1.0.0 (2020-07-17)
 
-
 ### Features
 
-* **crypto-store:** first stable version ([9e9417e](https://github.com/willgm/web-crypto-store/commit/9e9417eeeeda5eb2f32162e7d65ea9d7c32efb69))
+- **crypto-storage:** first stable version ([9e9417e](https://github.com/willgm/web-crypto-storage/commit/9e9417eeeeda5eb2f32162e7d65ea9d7c32efb69))